yolo-cls/yolo-cls/requirements.txt
ultralytics>=8.3.0,<9
fastapi>=0.111,<1
uvicorn[standard]>=0.30,<1
pillow>=10,<11
python-multipart
